+++ src/main/java/org/brapi/v2/api/ReferencesApiController.java
@@ -167,7 +167,7 @@ public class ReferencesApiController implements ReferencesApi {
 			pagination.setPageSize(result.getData().size());
 			pagination.setCurrentPage(0);
 			pagination.setTotalPages(1);
-			pagination.setTotalCount(result.getData().size());
+			pagination.setTotalCount((long) result.getData().size());
 			metadata.setPagination(pagination);
 	
 			rlr.setResult(result);
