+++ src/main/java/org/brapi/v2/api/ReferencesetsApiController.java
@@ -110,7 +110,7 @@ public class ReferencesetsApiController implements ReferencesetsApi {
 			pagination.setPageSize(result.getData().size());
 			pagination.setCurrentPage(0);
 			pagination.setTotalPages(1);
-			pagination.setTotalCount(result.getData().size());
+			pagination.setTotalCount((long) result.getData().size());
 			metadata.setPagination(pagination);
 			rlr.setMetadata(metadata);
